pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "echoharvester"
version = "0.1.0"
description = "Korean ASR Training Data Pipeline from YouTube and Local Media"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "EchoHarvester Team"}
]
keywords = ["asr", "speech-recognition", "korean", "youtube", "data-pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "yt-dlp>=2024.1.1",
    "ffmpeg-python>=0.2.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "webvtt-py>=0.4.6",
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "jiwer>=3.0.0",
    "lhotse>=1.20.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "tqdm>=4.66.0",
    "aiofiles>=23.2.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
echoharvester = "echoharvester.main:main"

[project.urls]
Homepage = "https://github.com/fitz04/EchoHarvester"
Repository = "https://github.com/fitz04/EchoHarvester"

[tool.setuptools.packages.find]
where = ["."]
include = ["echoharvester*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
